Minimum changes required to normalize paths
When binding a mounted path from a Linux host OS to a Windows OS, a mixture of slashes and backslashes resulted. As a result, the script was failing to recognize valid .json sidecar files. These paths have been normalized to slashes in this commit.

Sample YAML and ENV files have also been added.